+++ src/main/java/faang/school/accountservice/service/TariffService.java
@@ -2,10 +2,12 @@ package faang.school.accountservice.service;
 
 import faang.school.accountservice.model.dto.TariffDto;
 
+import java.math.BigDecimal;
+
 public interface TariffService {
     TariffDto createTariff(TariffDto tariffDto);
 
-    TariffDto updateTariff(Long id, Double rate);
+    TariffDto updateTariff(Long id, BigDecimal rate);
 
     TariffDto getTariff(Long id);
 }
